[Fixture] 修改ssh terminal, 支持request, g, client_channel上下文
