+++ .gitignore
@@ -1,5 +1,6 @@
 .idea
 *.pyc
 *.pyo
-.secret_key
+.access_key
+*.log
 logs/*
